Move NullUploadRunner to shared testlib, since it is used in two test modules. Add some extra unittest assert helper methods.
