+++ aodncore/testlib/__init__.py
@@ -1,12 +1,14 @@
 from .basetest import BaseTestCase
 from .dummyhandler import DummyHandler
 from .handlertest import HandlerTestCase
-from .testutil import (dest_path_testing, get_nonexistent_path, get_test_config, make_test_file, make_zip, mock)
+from .testutil import (NullUploadRunner, dest_path_testing, get_nonexistent_path, get_test_config, make_test_file,
+                       make_zip, mock)
 
 __all__ = [
     'BaseTestCase',
     'DummyHandler',
     'HandlerTestCase',
+    'NullUploadRunner',
     'dest_path_testing',
     'get_nonexistent_path',
     'get_test_config',
